Merged back from the trunk:
Alpha channel is now preserved when rendering PNGs from a PNG original with gd (not netpbm).
Persistent file upload image previews now use the same format as the original, taking advantage of the above.
We still plan to implement a "use netpbm only when images are big" mode, which should be the best of all possible worlds



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.3@543 c89fab08-2f9b-42bf-8176-bd3ad6bee547