New 'add' option to aMultipleSelect provides a straightforward way to allow admin users to add new categories, etc. while adding media and the like. Works well, needs styling, not yet integrated into the obvious places to use the feature such as media categories management.
If the original widget is named foo[categories][], the new items will submit as foo[categories_add][]. In general the name for the added categories is automatically chosen to parallel the name for the original widget in a sensible way. The added categories are of course submitted by name, not id, since they don't have any ids yet. 

The JavaScript does not allow the user to submit a blank value or a duplicate value, but obviously you must validate against both on the server side as it is trivial to spoof this and ignore the javascript code. <-- important



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@774 c89fab08-2f9b-42bf-8176-bd3ad6bee547