YOU MUST RUN ./symfony apostrophe:migrate AFTER THIS UPDATE (if you don't use MySQL, refer to the source of that task for commands you may need to adapt to update your schema, or do ./symfony doctrine:build-sql and look for the new a_redirect table)
A big change: when you rename a page, its slug is automatically updated as well. When you rename a page, edit its slug or move it via "reorganize," the slugs of its descendant pages are automatically updated accordingly. In the case of re-titling a page or editing its slug manually, descendants are only affected if their current slug begins with the old slug of the page you're editing. For reorganize the change always takes place in descendant pages.

When the slug of a page changes for any reason, a soft redirect is always established from the old URL. This redirect is sent to the browser only if there is no new page or engine at that URL. So if you replace a page with a new one, the new page will not be affected by the redirect. The soft redirects are always "301 Moved Permanently," so they are SEO-friendly.

In short: renames, slug edits and redirects are now magical. In the good iPhoneish way, not the scary unpredictable way.

Also: the rename form is now properly bound and validated. The apostrophe:migrate task, which you must run after updating to this revision, now updates the database schema as described in last night's email to the apostrophenow mailing list. Zend Search doesn't try to use the UTF-8 lexer/tokenizer unless iconv is actually available. A new validator has been created for slugs. The I18N-capable aTools::slugify method now does a better job, removing leading and trailing dashes and folding consecutive dashes.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@846 c89fab08-2f9b-42bf-8176-bd3ad6bee547