Allow 1000 character slugs to handle importing existing sites with deep structure. That's the practical limit with MySQL. Carefully specify the index length so MySQL actually pays attention to this.
Warning comments before various API methods in aPage explaining the need to get the page properly with retrievePageBySlugWithSlots() first. TODO: document this entire subject in the manual.

One can now specify whether the slot should go to the top or bottom of the area when adding a slot with newAreaVersion (TODO: document this method generally and this option particularly)

Correctly check for an instance before using sfContext to check privileges

Fixed a warning in apostrophe:deploy

Removed some obsolete, noisy logging calls



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@969 c89fab08-2f9b-42bf-8176-bd3ad6bee547