new apostrophe:import-files task pulls in JPEGs, GIFs, PNGs and PDFs in the specified folder, which defaults to web/uploads/media_import. Then it REMOVES those files from that folder.
Arrange FTP or SFTP access for the client to drop off the files, or upload their CDs yourself, and run the task... you're good to go.

TODO: import metadata. For now, the filename is slugified as a title and slug, and any subdirectory names in the path are made into tags. (This is a surprisingly common strategy in shops that aren't XMP-aware.) We'll add XMP import soon!



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@1254 c89fab08-2f9b-42bf-8176-bd3ad6bee547