Improved handling of virtual pages in search results. If your virtual page slug looks like a Symfony URL:
@foo?arg=val&... 

Or:

module/action?arg=val&...

Then it is handled as such. This allows you to create virtual pages that have a permalink for purposes of search results. We use this in our new blog plugin, in which blog post content is saved in ordinary Apostorphe slots in a virtual page. The slug of that virtual page is a Symfony route for an action that redirects you to a nice landing page for that blog post on the most appropriate public blog engine page.

If your virtual page should not appear in search results, do not begin it with a @ or include a / inside it.

Slugs with / at the beginning are, of course, normal CMS pages and are handled as such with aTools::urlForPage.

Suggested fixtures change: removed a_blog_admin permission, since we just look for admin.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@1288 c89fab08-2f9b-42bf-8176-bd3ad6bee547