getAncestorsInfo now has an optional $livingOnly flag to return only ancestors that are not archived.
getAccordionInfo no longer returns archived ancestors when $livingOnly is true. This is a change of philosophy however those with editing privileges are usually permitted to see navigation with $livingOnly set to false anyway, and those without really shouldn't see archived things in the tree above them. So this allows us to accommodate the "unpublish an ancestor to create pages that don't show up in navigation until you're there" trick better



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@1756 c89fab08-2f9b-42bf-8176-bd3ad6bee547