Big (yet backwards compatible) changes:
1. From now on the goal is to eliminate script blocks and old-school jq_do_something helpers from our PHP code.

As a down payment on this I migrated the slideshow javascript (pretty big) and the search field labeling javascript (pretty tiny) to show how it's done.

All JavaScript code that is specific to Apostrophe should live in methods of the 'apostrophe' object, which we have had for a long time. It lives in a.js. To add another method there:

this.methodname = function(options) {
  Code goes here
};

Now you can call apostrophe.methodname(options) in JavaScript. Nicely namespaced.

2. But how do you invoke one of these methods without a script block? You use the new a_js_call helper:

<?php a_js_call('apostrophe.methodname', array('title' => 'this is one parameter', 'count' => 10)) ?>

This does several good things:

1. It automatically and safely json_encodes the array of parameters as a JavaScript object, which shows up as the one and only parameter to the JavaScript method. No escaping gymnastics and bugs all over the place our PHP code.

2. Your javascript method always receives a single 'options' parameter, so you wind up referring to options like this:

options['title']

Which is way more flexible than comma-separated arguments and you always wind up adding an options parameter eventually anyway when you realize you have too many arguments.

3. The JS call is deferred until a_include_js_calls() is invoked at the end of the body in layout.php, or at the end of the AJAX slot update partial. When you call a_include_js_calls(), a DOMready javascript block appears with all of the calls that were made, in order, in that single block. This is way, way more efficient on the browser side - inline script elements block progressive loading of images etc. until they are done running - doing it this way forces it to happen at the end on domready for best performance and reinforces the idea that our pages should be prepared to render without a FOUC without the help of JavaScript (if something is totally not ready without JavaScript, hide it in CSS and show it when the JS runs, but better yet avoid that).

The goal now is to burn through apostrophePlugin and apostropheBlogPlugin looking for script blocks and JavaScript helper calls, migrate the code to reusable, parameterized methods in apostrophe or perhaps a new aBlog object in an aBlog.js file, and replace all of the invocations with nice clean calls to the a_js_call() helper.

* * *

I also fixed a few seemingly old bugs in the next/previous buttons in the slideshow partial.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@1964 c89fab08-2f9b-42bf-8176-bd3ad6bee547