Apostrophe now supports granting groups editor and manager privileges at any particular point in the page tree, building on our strong support for granting privileges to individuals.
This closes #520

I need to do some tuning to cut down on the number of SQL queries involved.

Add the new 'editor' permission to the GROUP (not the individual) in order to make it possible for that group to be granted privileges. This is analogous to adding individuals to the 'editor' group to allow them to be granted privileges singly.

Be sure to run the apostrophe:migrate task.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2016 c89fab08-2f9b-42bf-8176-bd3ad6bee547