Wrapped tree lock calls around page creation and deletion to address the fact that Doctrine doesn't seem to have any concurrency locks for nested set operations. We had previously locked reorganize operations for similar reasons but did not realize that the fundamental insert and delete operations did not have locks either (transactions do not address the same issue).
On a related subject, the repair-tree task has been overhauled. All variants of the task now use PDO to avoid running out of memory on big page trees (they are very fast now). There is now a method option which can be set to list or slug. The list option (tested) doesn't reorganize a messed-up page tree but it does correct any errors in lft and rgt values such that it is now safe to manually reorganize it. The slug option infers the page tree from page slugs; this discards order of pages at the same level and doesn't work well if you heavily edit your slugs. A third approach is to specify the csv option, which should be set to a file containing a CSV dump with id, lft, rgt and level values (in that order, with no header) from a known-good database. Pages that did not exist in that good database become archived children of the homepage for easy cleanup in 'reorganize.' See the verbose help for the task for an example of how to create such a CSV file from a known-good backup.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2030 c89fab08-2f9b-42bf-8176-bd3ad6bee547