I updated aInputSelfLabel with a slight behavioral change.
On focus it places the cursor at the front of the input

On keydown it clears out the label, displaying what you type.

added a new jq function to the bottom of aControls called aSetCursorPosition. I found a piece of it on the internet and modified it to work with out stuff



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2051 c89fab08-2f9b-42bf-8176-bd3ad6bee547