Edit view forms are now AJAX loaded when the edit button is clicked for vastly better performance in browsers that are adversely affected by too many FCK instances, also faster and lighter page loads when logged in.
This is based on my original ajaxeditview feature branch. I had talked about reworking this change to apply to FCK only, but on closer inspection the merge was easy to do and along the easiest cutting edge by a long shot.

The only code change necessary: if you have written a custom slot, and you are including simpleEditButton rather than simpleEditWithVariants, then you need to start passing 'slot' => $slot as an additional option to that partial. 

You'll get a warning message to that effect (and your slot will not be editable) if you have not made this change yet.

Note that this does apply to slots created by the apostrophe:generate-slot-type task.

If you are using simpleEditWithVariants then you won't need to make any changes since that partial has been upgraded for you.

If you are overriding _slot, _simpleEditButton etc. you will need to make changes, but that is not surprising.

Apostrophe's standard slots are already upgraded. 

Closes [1375]



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2139 c89fab08-2f9b-42bf-8176-bd3ad6bee547