Fixed a longstanding subtle issue probably present in most PHP applications:
If you try to upload something that exceeds your php.ini post_max_size setting, the entire POST array is empty, leading to confusing results (in Apostrophe's case, missing CSRF tokens, leading you to get stuck and not being able to ever complete the PDF or image upload after that). 

I addressed this by checking for a POST request with an empty POST array and treating it as a GET request, with the addition of an error message that includes post_max_size.

This should fix the problem Dana ran into today on PI. John should update and deploy.

I need to merge this to the trunk.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.4@2194 c89fab08-2f9b-42bf-8176-bd3ad6bee547