Closes #593, significantly addresses #573 by merging the page settings and add page forms.
"Page Settings" and "Add Page" both open a full featured form that allows you to, for instance, create a blog engine page locked to the "fruit" category and allow only appropriate people access all in one pass. No more creating a temporary "normal" page on your way to doing what you really intended.

Page slugs can be customized immediately at the time pages are created. This is an important feature for clients who frequently want to create a shorter permalink that does not reflect the page tree.

The two use the same Symfony form class, and they share a template, settingsSuccess.php. When editing or overriding that template you can easily check whether you are dealing with a new page or not by checking $page->isNew(). This is already being done to remove the delete button for a new page and label the save button appropriately.

Beginning in 1.5 engine settings forms must validate when the page object is new. However the page will be saved first (and your form reinstantiated and rebound), so the save method of your engine settings form will know the title, etc. of the page.

Having only one implementation of page settings will improve maintainability as well as improving workflow for end users.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2221 c89fab08-2f9b-42bf-8176-bd3ad6bee547