The new view permissions are roughed in and working.
Permissions need styling, really a lot.

Right now only admins can change or observe the settings of permissions for view or edit. That got us there today, but we can open it up a little later as discussed - editors could have control of view permissions (although if you can edit you should always be able to view, so you can't lock out your peers). And we could have some kind of read-only view of the edit permissions for managers and editors if that's really appropriate (let's discuss that)



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2365 c89fab08-2f9b-42bf-8176-bd3ad6bee547