Fixed a lot of search stuff, providing the underpinnings for #670 and #623 and also other good things.
It turns out that if you try to add a field in Zend Lucene as both an indexed field and a stored field, the last one wins, which is why all of my previous attempts to make fields other than the full text individually searchable didn't work out.

(We could use Doctrine "Text" fields, which are both indexed and stored, but we need to convert to lowercase when indexing, which would clobber our presentation of search results, so we don't do that.)

We now store fields with a "_stored" suffix to avoid any possibility of collision with an indexed field of the same name.

Overrides of the a/search action will be affected, but templates should not be.

For purposes of #670 the big benefit here is that you can now do a search like this one:

slug:@a_blog_search_redirect AND dwarves

To find only blog posts that mention dwarves. Regular pages mentioning dwarves will NOT be returned.

I should be able to leverage this to implement real blog search both for adding posts to a blog slot (#623) and for generally navigating the blog (#670).

There are other benefits too. You can now search on titles only:

title:monkeys
title:"monkey mittens"

You can also search on tags or metadescription.

Also, titles and tags and metadescriptions should now be correctly boosted since we're successfully indexing them as separate fields at this point.

YOU NEED TO DO an apostrophe:rebuild-search-index after this one.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2506 c89fab08-2f9b-42bf-8176-bd3ad6bee547