Implemented #670. There is now blog search in the blog sidebar, which is aware of both publication dates and the category restrictions of the blog engine page you are searching on.
*** YOU NEED TO:

./symfony doctrine:build --all-classes
./symfony apostrophe:migrate
./symfony apostrophe:rebuild-search-index

After updating to this revision ***

The UI for search needs styling, I borrowed the full site search markup pretty carefully but it's not happy in the sidebar yet. There *is* a search field that works, it's underneath the area it seems like it should be in.

I did not implement events search. Let's wait until the dust has settled with blog search.

I am not 100% thrilled because I'm filtering categories at the Doctrine level after Lucene search is done, which is not good if you already restricted Lucene to 1000 matches, there are 2000 blog posts in your system that match your terms better, and only 10 are in your category. I got much better at handling Zend search queries in complex ways but backed some of that out because I wasn't expressing it very cleanly. I want to look at that issue some more when I get a chance but this should work well for most purposes.

There is also an as yet untested and possibly busted upgrade to aMultipleSelect here, it's working for blog but either it's not working for page settings or page settings has another bug. I'll get that wrangled properly as part of #623 (which also needs all the work done here).



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/trunk@2514 c89fab08-2f9b-42bf-8176-bd3ad6bee547