Fixes #715. Dan's recent refactoring (which I reviewed and approved) of queryWithSlots switched to using 'andWhere' calls to add the criteria to select relevant slots. The problem is that this acts as an implicit INNER JOIN, wrecking our carefully constructed LEFT JOIN, the purpose of which is to make sure we always get back a page if it exists, even if it has no slots yet.
This led to a proliferation of pages with the "global" slug on each page refresh.

Oops.

I brought back the use of WITH clauses in the leftJoin calls but did it in such a way that it should be just as reusable as Dan's code.

Thanks to Geoff Hammond for reporting the problem.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@2641 c89fab08-2f9b-42bf-8176-bd3ad6bee547