ATTN: I overrided a portion of FCK that pertains to the Right-Click menu for UL and OL elements.
FCK was using old html attributes for specifying the list-type and this is bad, wrong, and just functionally doesn't work when we use things like reset stylesheets.

So I updated FCK to instead write inline styles for list-style-type.

The only caveat to this is that you will have to explicitly allow the style attribute on OL and UL and then explicitly allow list-style-type property in the project's app.yml.

I will not be forward porting this to 1.5 because we have different ideas for how to handle this moving foward.




git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.4@2678 c89fab08-2f9b-42bf-8176-bd3ad6bee547