Implemented #734: uploaded thumbnails are once again allowed (though not mandatory) for videos from services we don't support directly. In addition, since it was simplest to permit it, you can replace the thumbnail from YouTube with a custom thumbnail image file if you prefer.
Also, the same _edit partial is used for all annotation situations, including annotating new videos. This makes maintenance much easier.

There is still a fair amount of special case code at the action level, but most of it is necessary to deal with things like pulling thumbnails from services. We're abstracting that better, preloading the persistent file widget rather than special-casing it in the form



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@2865 c89fab08-2f9b-42bf-8176-bd3ad6bee547