Fixed #141: cancel the media selection if the user browses away from the media repository by following a link. I've done this as conservatively as I can to avoid breaking stuff. There is a synchronous AJAX request to cancel the selection, then the link operates as it normally would. This should avoid breaking things that have other, equally or more important fancified javascript handlers on them.
IMPORTANT: if you are creating a button or link that will just be a hook for media related javascript please *don't* write href="/#" as the leading / means my code can't detect this as an "inside the media repository, don't cancel" link. Instead write href="#" so it is clear that this is "within the page" and not a link to the homepage. I had trouble with a couple of these and wound up grepping for them.

The "cancel selection, then let the link operate normally" behavior assumes that links starting with '#' and links containing '/admin/media' anywhere inside them are local to the media repository and should not cancel the selection. Anything else cancels the selection before operating normally.

"Open link in new tab" does not trigger javascript at all, so it is not affected. This is good because it means you can open a tab to go manage tags or something without busting your selection, but I can't really take credit for it.

* * *

In the long term the whole business of selecting media should probably become a roll-down, "90% of the page" experience inside the page you came from, which we can think about for 1.6, certainly for 2.0.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@2958 c89fab08-2f9b-42bf-8176-bd3ad6bee547