Fixed #822: later pages of YouTube search results mysteriously returning nothing.
YouTube reports giant #s of results but actually caps results at 1,000 (even if you are paginating nicely) and will also refuse to give you any results if you ask for 9 more items starting with item 1000 (you must ask only for that one item).

So we no longer claim more than 1,000 results, and we obey the right protocol to get the lone result on the 112th page (the 1,000th result).

* * *

Implemented #820: display helpful message to the admin when some embedded services are not configured yet. There is now a helpful message linking to a documentation page explaining how to configure the Vimeo embed service. This message goes away once you take care of it. Only admins see the message. This is implemented in an extensible way so future embed services can also have a page somewhere that documents how to configure them when needed.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@2989 c89fab08-2f9b-42bf-8176-bd3ad6bee547