Resolved #832 (Apostrophe reacts unreasonably when a video is not permitted to be played in your country) and #776 (videos should be replaceable) with a significant refactoring of our embedded media support. As it turns out #832 was a symptom of a bigger problem (we had no good way to handle validation errors on embed codes) which was a symptom of an even bigger cowardice on my part (I was afraid of #776). I got it wrangled.
There is now always an "Embed Code or URL" field when you edit a video, it works just like the field you paste the code or URL into initially, and displays the URL in preference to the embed code for an existing item if we have direct support for that service (like YouTube). 

This means that you can replace a YouTube video with a Viddler video or vice versa, even though one is native and the other is not, and media slots that use it will just pick up the change like they would if you replaced a file. 

This is a big usability and code maintainability improvement, it was well worth the time.

NOTE: the thumbnail does NOT change, and the metadata does NOT change, if you replace one youtube video with another. This is in keeping with the way we approach metadata everywhere else: once you save it the first time it is assumed that you have hand-tweaked it, and if you replace the file or the video you do NOT want your metadata rewritten for you. 

I think this is correct for other metadata but the thumbnail part is debatable. Maybe you replaced the original thumbnail with something you designed specifically. Then again maybe you didn't and you'd like a relevant automatic thumbnail from YouTube. Designing a UI to let the user choose to refresh the thumbnail from the service is tricky though because that UI would have to jump into view when you put in an embed code that we natively support, and not before. I think for 1.5 we leave it as it is



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@3017 c89fab08-2f9b-42bf-8176-bd3ad6bee547