Implemented #484: all slashes in tags are converted to dashes. This avoids a serious issue with both major web servers which are typically configured to refuse to permit an escaped slash in a URL, ever. This is done in apostrophe:migrate and also for all new saves of tags.
Also, we were converting tags to lowercase in some places but not others. After consultation I've determined that we *want* mixed case in tags, although we compare tags in a case-insensitive way (if we have the tag apple and the tag Apple comes along, that's the same tag). The tag admin can then be used to fix the case of any proper name tags that started out lowercase, etc. 

There was some discussion of not showing spaces as + signs in permalinks to tags, or perhaps having actual slugs for tags, however since this is deep in the guts of Symfony, the webserver and the tag plugin (which does not have slugs for tags) this is not practical for 1.5



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@3175 c89fab08-2f9b-42bf-8176-bd3ad6bee547