The simpleEditButton and the Save/Cancel javascript that lives in the _slot used .parent() and .children() for traversing.
This is problematic when you want to customize the markup in the area or slot at the project level because all of the buttons events break due to scoping.

I updated the javascript to use .find() when traversing down, and .closest() when traversing up, so that it is both more flexible and more specific.

I believe this is still an issue in A1.5 and I will take a look at addressing this over there as well.




git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.4@3277 c89fab08-2f9b-42bf-8176-bd3ad6bee547