The new aCacheFilter '''drastically speeds up high-traffic Apostrophe sites''' by caching all content for GET requests for five minutes for logged-out users. This is great, but it does require that your editors clearly understand that their edits will not be seen by logged out users for up to five minutes. This simple strategy avoids many, many deeply complex issues with cache invalidation. But since it requires user education and is not necessary on low-traffic sites we do not enable it by default in our sandbox project.
aCacheFilter does not yield as big a speedup as a static cache, but also does not have
cache invalidation problems once the user logs into the same page they just saw as a logged-out user.
It is a caching solution for logged-out users, designed to play nicely with Apostrophe's in-context 
editing philosophy for logged-in users.

(Note that you will not see a performance benefit if no one else has accessed the page
in question in the past five minutes. If your site does not receive high traffic
you probably do not need this feature.)

Activate aCacheFilter in `apps/frontend/config/filters.yml` like so:

{{{
rendering: ~
security:  ~

# insert your own filters here
aCache:
  class: aCacheFilter
  
cache:     ~
execution: ~
}}}

In app.yml you can adjust the cache lifetime (which defaults to 300 seconds), the Symfony cache class used (which defaults to sfFileCache), and the options passed to the cache class (which default to a reasonable setting for the cache folder name). Note that this means you can switch to sfAPCCache or or sfMemcacheCache if you prefer them:

{{{
all:
  a:
    page_cache_lifetime: 300
    page_cache_class: sfFileCache
    page_cache_options:
      cache_dir: /your/project/root/dir/data/a_writable/a_page_cache
}}}

One big gotcha: sometimes logged-out users do things that change content 
generated by GET requests. For instance, you might have code that implements
blog comments for users without accounts. They add a comment via a POST request
and you redirect them (GET request) to the blog post page they have already 
seen. You try to use a flash attribute to add a "thanks for your comment"
message but it never shows up because the blog page is coming from the cache.

The solution is to set the aCacheInvalid flash attribute to true for the user
in your "executeComment" action:
{{{
$this->getUser()->setFlashAttribute('aCacheInvalid', true);
}}}

This will prevent the cache from being consulted on the next request
for this user.

You can also block the cache entirely for the remainder of a user's session
with the regular setAttribute method:

{{{
$this->getUser()->setAttribute('aCacheInvalid', true);
}}}

The latter fix is better if you want the user to see the results of a
change you have permitted them to make consistently as they move through 
previously visited pages on the site.

This is useful when you have users who are not really logged in 
according to Symfony (isAuthenticated() is false) but they are still
doing interactive things that alter enough content that the cache becomes
confusing for them. 

Our advice is to do this generously whenever you think
it's needed since the main purpose of the cache is to handle the bulk of
casual, logged-out site visitors efficiently. 



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@3737 c89fab08-2f9b-42bf-8176-bd3ad6bee547