* Notify an a.afterTreeMove event after moving a page in the tree via the reorganize feature. This allows FM to force particular template settings for certain neighborhoods * Set a scroll_spd that is not crazy-slow
Note that to actually get scrollability in the version of jstree you're using the tree element itself must be overflow: scroll. You can't scroll the main document automatically when dragging, you have to have a scrollbar on the tree. We might want to do this in general, I certainly had to do it for FM	 
 


git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4005 c89fab08-2f9b-42bf-8176-bd3ad6bee547