#1147 implemented: introducing aSearchService, a better alternative to Zend Lucene search in Apostrophe, and aMysqlSearch, a simple sample implementation with no external dependencies, good performance and complete integration with Doctrine queries.
For backwards compatibility and stability, Zend Lucene search is still the default behavior. Nothing changes unless you expressly activate a subclass of aSearchService in app.yml, for example:

all:
  a:
    # Found in apostropheMysqlSearchPlugin
    search_service_class: aMysqlSearch

Once you switch to a search service, all searches go through the interface provided by the public methods of the aSearchService class.

The most important is aSearchService::addSearchToQuery(), which adds a search to a Doctrine query. 

** The aMysqlSearch implementation does this entirely using proper Doctrine joins **, with major performance benefits. However it is possible to create alternative implementations which do it by searching an external search engine and then using addWhereIn() and FIELD() to merge the results into a Doctrine query. This means that third parties can do Apache Solr integration, if they really want to.

Most usages of the old aZendSearch API still work when a search service is activated. They use the search service to provide forwards compatibility. Some specialized cases, such as working with the methods of Zend search hit objects and documents directly, will need to be rewritten (see BaseaActions::search for how we did this for the main search in Apostrophe, which is now fully native aSearchService goodness).

** Be aware that overriding aActions::addSearchResults() is not supported with search services. ** It was a source of serious performance issues and one of the reasons why proper offset/limit clauses could not be used with search results. However there are two alternatives available to you: (1) override a/searchBefore and/or a/searchAfter to create search sidebars that provide teaser search results for other data types that users can click on to get more information (recommended). (2) Mirror your content to an Apostrophe virtual page with a slug that is a valid Symfony URL which displays (or redirects to) the item in question. Search results recognize such virtual pages and include them in search results. This is how blog articles are included in the main Apostrophe search results.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4071 c89fab08-2f9b-42bf-8176-bd3ad6bee547