Implemented #1170: make Apostrophe load-balancing and cloud-storage friendly.
See http://trac.apostrophenow.org/wiki/ManualScaling for a comprehensive discussion of how to use these new features.

The default behavior of the sandbox has not changed with regard to documented features.

Be aware that if you are using the undocumented "feature" that you can generate a media image in any size by constructing a URL with width, height, etc. values in it, this is no longer available. This change was made because that "feature" allows for a DOS attack to be carried out by peppering Apostrophe with URLs for random image sizes. With cloud hosting this is especially dangerous because it also incurs charges. 

The right way to do that is now to call the getScaledUrl() method on an aMediaItem object. This generates an authorization allowing that image to be rendered when the img src request arrives. If you are using cloud file hosting the image is then kicked upstairs to the cloud and the img src request is redirected.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4201 c89fab08-2f9b-42bf-8176-bd3ad6bee547