The app_a_less_import_directory option just got a lot more useful.
On certain projects we have trouble because we want to be able to @import file c from file b which is imported from file a, but all import paths are relative to file a, which typically breaks the @import c; statement in file b because that statement is expecting a path relative to file b.

I didn't try to change this fundamental behavior in lessc, because that would be a bc break.

However, lessc does support specifying an array of import directories rather than just one, and it takes the first match it finds when seeking a particular file. 

This is enough to solve the problem already - you could specify an array of fallbacks in app.yml, separately for dev, staging and prod, with full paths.

But it makes more sense to allow the use of SF_PLUGINS_DIR and SF_WEB_DIR to refer to the plugins folder and the project folder without entering full paths for specific servers. And a similar way to refer to the directory that file a was found in is nice to have. So I implemented SF_PLUGINS_DIR, SF_WEB_DIR, and SELF_DIR.

This way you can easily specify several levels of fallback in app.yml: "look at project level, then in this overriding plugin, then in apostrophePlugin, and then in the folder the darn less file came in," or similar.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4339 c89fab08-2f9b-42bf-8176-bd3ad6bee547