Added a few 24-bit PNGs for old IE support. For modern browsers we have RGBA.
A few places where we use the a-rgba helper, I replaced the solid color fallback with the 24-bit PNG image

Removed a-ie.less because we don't / can't compile the ie conditional stylesheet, the regular CSS version is what is included as a comment in layout.php

TODO: I am going to make the layout in apostrophePlugin use the HTML5 doctype based on the HTML5 Boilerplate. That allows us to use the nifty <HTML> tag classes that we are accustomed to using with the apostropheHTML5Plugin. This makes maintaining IE style fixes a breeze.




git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4364 c89fab08-2f9b-42bf-8176-bd3ad6bee547