Remember clearing your browser cache after deployment and hoping your users would do the same? Yeah, that's no longer a thing. Asset URLs are always unique to the latest deployment on staging and production, assuming you have our minifier turned on.
For this to work properly you have to set it up in app.yml too:

a:
  <?php if (file_exists(dirname(__FILE__) . '/asset_generation.yml')) require dirname(__FILE__) . '/asset_generation.yml' ?>

This takes care of supplying a unique prefix for all minified asset groups that changes every time the apostrophe:bump-asset-generation task is run. And apostrophe:deploy runs that task remotely, which means it should be automatic on all staging and production deployments.

Thanks to Jeremy for nagging me to do this a long time ago. I finally got around to it. Loading the asset generation via app.yml in this way means there is no performance hit as there would be if we loaded it from a file or the cache on every page load.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4783 c89fab08-2f9b-42bf-8176-bd3ad6bee547