Want to toggle open a filter exactly the way the list of tags toggles open in the blog? Do something like this:
<div class="issues">
	<h4 data-toggles=".issues .a-filter-options" class="a-slide-toggle">Issues</h4>
	<div class="a-filter-options a-slide-toggled">
	  ... Filters go here ...
	</div>
</div>

The important part is that you give your label the a-slide-toggle class and set the data-toggles attribute to a selector that finds the thing to be toggled.

This capability is part of apostrophe.smartCSS, so it is always available, you don't have to specially enable it, just use the classes and you're good to go.

This was refactored from the blog plugin.

The blog filter support in the entities plugin now uses it.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4874 c89fab08-2f9b-42bf-8176-bd3ad6bee547