aTools::getUniqueName takes a user or an array representing a user and returns a unique name for that user suitable for use in a select element. The default behavior is:
Firstname Lastname (username)

You can override this in your project-level aTools:: class.

Right now this method is used to populate the "author" and "editors" dropdowns in the blog admin, which yields a very large performance improvement over using object hydration in these forms when there are many users.

(In principle sfGuardUserTable is the right place for this, but Apostrophe doesn't own that.)



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4929 c89fab08-2f9b-42bf-8176-bd3ad6bee547