The a_tab helper returns a tab compatible with the markup used by Apostrophe's standard navigation. This is useful when your site is not dominated by Apostrophe pages and you wish to implement your own tabs in the same style.
This helper is similar to link_to in that the URL parameter
is passed through url_for, so @my_route is automatically
converted to a final URL for you.

The following conveniences are implemented to streamline the
effort involved:

The 'first' class is automatically set on the first tab that
is returned during this PHP request as a whole. You can override
this behavior by passing the 'first' option (true or false).

The '' class is automatically set on the tab if the
URL generated by the $url parameter matches the current
request URL or appears at the start of the current request
URL, followed by a slash. You can override this behavior by
passing the 'current' option (true or false).

Additional options:

The 'class' option is supported and is applied
to the <li> tag for the tab.

The 'open' option, if true, prevents </li> from being
added automatically. This is useful if you are implementing
dropdown menus via nested <ul>s.

The 'link' option, when present, contains attributes to be passed
to the 'a' element inside the 'li' element.

Note that overrides of 'class' for either the <li> or the <a> will
always be preceded by the standard classes for our tabs.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4932 c89fab08-2f9b-42bf-8176-bd3ad6bee547