Doctrine::getTable('aPage')->addPage($parent, $title, $options) adds a new page, copying permissions from the parent as you would expect.
Returns a page object. You can then call newAreaVersion or whatever you'd like on that object to add slot content.

IT IS YOUR RESPONSIBILITY to call aTools::lock('tree') at the start of your  ENTIRE action and aTools::unlock() at the end. OTHERWISE THE ENTIRE PAGE TREE CAN BE LOST if a race condition occurs. I can't do it for you here because you need to do it before you retrieve the parent page.

THE TITLE IS ASSUMED TO BE PLAINTEXT, so this method entity-escapes it for you.

Currently you can override the following options:

template, slug



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4978 c89fab08-2f9b-42bf-8176-bd3ad6bee547